requirements.txt
# AI Prompt Manager Dependencies

# Core dependencies
requests>=2.25.0
python-dotenv>=0.19.0

# CLI enhancements
pyperclip>=1.8.2

# Optional but recommended for better user experience
# colorama>=0.4.4  # Cross-platform colored terminal text
# rich>=10.0.0     # Enhanced CLI formatting (uncomment if needed)

# Development dependencies (uncomment for development)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.8.0
